requirements.txt
openai
python-dotenv
pylint
flake8
black
pre-commit
